+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * perf: Move instantiable check to `ObjectMappingResolver`.
 * refactor: Do not use `array_intersect` to determine object mapping.
 * dx(`ObjectMapping`): Make it mutable.
+* refactor(`ObjectToObjectMetadata`): Class & namespace renames for clarity.
 
 ## 0.5.18